+++ Backend/controllers/web.controller.mjs
@@ -84,7 +84,19 @@ export const allApplications = async (req, res) => {
 
     res.status(200).json(response);
   } catch (error) {
-    console.log(`error in employeeApplications ${error.message}`);
+    console.log(`error in allApplications ${error.message}`);
+    return res.status(500).json({ error: `internal server error on web controller` });
+  }
+};
+
+export const allCompanyApplications = async (req, res) => {
+  try {
+    const response = await CompanyRequest.find().sort({ createdAt: -1 });
+    if (response.length === 0) res.status(404).json({ error: `No applications found` });
+
+    res.status(200).json(response);
+  } catch (error) {
+    console.log(`error in allCompanyApplications ${error.message}`);
     return res.status(500).json({ error: `internal server error on web controller` });
   }
 };
