+++ Backend/server.mjs
@@ -28,9 +28,11 @@ const PORT = process.env.PORT || 5000;
 
 app.use(
   cors({
+    origin: "http://localhost:2000",
     credentials: true,
   })
 );
+
 app.use(express.json({ limit: "50mb" })); //this can cause DOS atatacks bit needed to upload pdfs
 app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
