--- Frontend/Admin/src/Components/AdminDashboard/AddCompany.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const AddCompany = () => {
-  return <div>AddCompany</div>;
-};
-
-export default AddCompany;
