+++ Frontend/Admin/src/Components/AdminDashboard/AdminDashboard.jsx
@@ -9,19 +9,19 @@ const AdminDashboard = () => {
   return (
     <React.Fragment>
       <div className="grid grid-cols-6 grid-rows-10 sm:grid-rows-6 sm:grid-cols-9 gap-5 w-full h-screen">
-        <div className="col-span-6 row-span-4 sm:col-span-3 sm:row-span-3 overflow-hidden shadow-md rounded-2xl pb-5">
+        <div className="col-span-6 row-span-4 sm:col-span-3 sm:row-span-3 overflow-hidden rounded-2xl pb-5">
           <PieChart />
         </div>
 
-        <div className="col-span-6 row-span-2 sm:col-span-6 sm:row-span-2 shadow-md rounded-2xl">
+        <div className="col-span-6 row-span-2 sm:col-span-6 sm:row-span-2 rounded-2xl">
           <Summary />
         </div>
 
-        <div className="col-span-2 row-span-4 sm:col-span-3  sm:row-span-4  shadow-md rounded-2xl">
+        <div className="col-span-2 row-span-4 sm:col-span-3 sm:row-span-4 shadow-md rounded-2xl">
           <AdminWindow03 />
         </div>
 
-        <div className=" col-span-2 row-span-4 sm:col-span-3 sm:row-span-4 shadow-md rounded-2xl">
+        <div className=" col-span-2 row-span-4 sm:col-span-3 sm:row-span-4 rounded-2xl">
           <MiniNotificationWindow />
         </div>
