+++ Frontend/Admin/src/Components/AdminDashboard/Cards/DashboarsCards/Summary.jsx
@@ -26,19 +26,19 @@ const Summary = () => {
 
   return (
     <div className=" h-full  font-bold p-4 flex justify-around flex-wrap ">
-      <div className="flex flex-col justify-center items-center drop-shadow-lg  rounded-2xl">
+      <div className="flex flex-col justify-center items-center drop-shadow-  rounded-2xl">
         {fetching ? <div className="flex justify-center items-center  ">{<PuffLoader />}</div> : ""}
         <span className=" text-3xl">{employeeCount}</span>
         <span className=" text-xl">No of Employess</span>
       </div>
 
-      <div className="flex flex-col justify-center items-center drop-shadow-lg  rounded-2xl">
+      <div className="flex flex-col justify-center items-center drop-shadow-  rounded-2xl">
         {fetching ? <div className="flex justify-center items-center  ">{<PuffLoader />}</div> : ""}
         <span className=" text-3xl">{companyCount}</span>
         <span className=" text-xl">No of Companies</span>
       </div>
 
-      <div className="flex flex-col justify-center items-center drop-shadow-lg  rounded-2xl">
+      <div className="flex flex-col justify-center items-center drop-shadow-  rounded-2xl">
         {fetching ? <div className="flex justify-center items-center  ">{<PuffLoader />}</div> : ""}
         <span className=" text-3xl">{applicationCount}</span>
         <span className=" text-xl">New of Applications</span>
