+++ Frontend/Admin/src/context/CompanyContext.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const CompanyContext = createContext();
