+++ Frontend/Admin/src/hooks/useEmployeeContext.jsx
@@ -4,7 +4,7 @@ import { EmployeeContext } from "../context/EmployeeContext";
 export const useEmployeeContext = () => {
   const context = useContext(EmployeeContext);
   if (!context) {
-    throw Error("useWorkoutContext must be used inside an WorkoutContextProvider");
+    throw Error("useWorkoutContext must be used inside an EmployeeContextProvider");
   }
   return context;
 };
